+++ app/src/main/java/com/king/mlkit/vision/app/Function.kt
@@ -24,27 +24,7 @@ import com.king.camera.scan.util.LogUtils
 /**
  * @author <a href="mailto:jenly1314@gmail.com">Jenly</a>
  */
-fun Bitmap.drawBitmap(block: (canvas: Canvas,paint: Paint) -> Unit): Bitmap {
-    var result = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888)
-    try {
-        val canvas = Canvas(result)
-        canvas.drawBitmap(this, 0f, 0f, null)
-        val paint = Paint()
-        paint.strokeWidth = 4f
-        paint.style = Paint.Style.FILL_AND_STROKE
-        paint.color = Color.RED
-
-        block(canvas,paint)
-
-        canvas.save()
-        canvas.restore()
-    } catch (e: Exception) {
-        LogUtils.w(e.message)
-    }
-    return result
-}
-
-fun Bitmap.drawRect(block: (canvas: Canvas,paint: Paint) -> Unit): Bitmap {
+fun Bitmap.drawRect(block: (canvas: Canvas, paint: Paint) -> Unit): Bitmap {
     var result = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888)
     try {
         val canvas = Canvas(result)