+++ app/src/main/java/com/king/mlkit/vision/app/MainActivity.kt
@@ -17,7 +17,10 @@ package com.king.mlkit.vision.app
 
 import android.Manifest
 import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.ImageDecoder
 import android.net.Uri
+import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
 import android.view.View
@@ -28,7 +31,6 @@ import com.google.mlkit.vision.barcode.common.Barcode
 import com.king.app.dialog.AppDialog
 import com.king.app.dialog.AppDialogConfig
 import com.king.camera.scan.CameraScan
-import com.king.camera.scan.analyze.Analyzer
 import com.king.camera.scan.util.LogUtils
 import com.king.camera.scan.util.PermissionUtils
 import com.king.mlkit.vision.app.barcode.BarcodeScanningActivity
@@ -46,13 +48,22 @@ import com.king.mlkit.vision.app.segmentation.SelfieSegmentationActivity
 import com.king.mlkit.vision.app.text.TextRecognitionActivity
 import com.king.mlkit.vision.barcode.BarcodeDecoder
 
-
 /**
+ * 演示示例
+ *
  * @author <a href="mailto:jenly1314@gmail.com">Jenly</a>
  */
 class MainActivity : AppCompatActivity() {
 
-    var isQrCode = false
+    private var isQrCode = false
+
+    private var toast: Toast? = null
+
+    private fun showToast(text: String) {
+        toast?.cancel()
+        toast = Toast.makeText(this, text, Toast.LENGTH_SHORT)
+        toast?.show()
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -85,14 +96,14 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    fun getContext() = this
+    private fun getContext() = this
 
     /**
      * 扫描结果
      */
     private fun processScanResult(data: Intent?) {
         val text = CameraScan.parseScanResult(data)
-        Toast.makeText(this, text, Toast.LENGTH_SHORT).show()
+        showToast("$text")
     }
 
     /**
@@ -101,51 +112,56 @@ class MainActivity : AppCompatActivity() {
     private fun processPhoto(data: Uri?) {
         data?.let {
             try {
-                val srcBitmap = MediaStore.Images.Media.getBitmap(contentResolver, it)
-                BarcodeDecoder.process(srcBitmap, object : Analyzer.OnAnalyzeListener<List<Barcode>?> {
-                    override fun onSuccess(result: List<Barcode>) {
-                        if (result.isNotEmpty()) {
-                            val buffer = StringBuilder()
-                            // 识别成功，在图片上框出结果
-                            val bitmap = srcBitmap.drawRect { canvas, paint ->
-                                for ((index, barcode) in result.withIndex()) {
-                                    buffer.append("[$index] ").append(barcode.displayValue)
-                                        .append("\n")
-                                    barcode.boundingBox?.let { box ->
-                                        canvas.drawRect(box, paint)
-                                    }
+                BarcodeDecoder.process(
+                    BarcodeDecoder.fromFilePath(getContext(), it),
+                    // 如果指定具体的识别条码类型，速度会更快
+                    if (isQrCode) Barcode.FORMAT_QR_CODE else Barcode.FORMAT_ALL_FORMATS
+                ).addOnSuccessListener(this) { result ->
+                    if (result.isNotEmpty()) {
+                        val buffer = StringBuilder()
+                        // 成功；在图片上框出结果
+                        val bitmap = it.getBitmap().drawRect { canvas, paint ->
+                            for ((index, barcode) in result.withIndex()) {
+                                buffer.append("[$index] ").append(barcode.displayValue)
+                                    .append("\n")
+                                barcode.boundingBox?.let { box ->
+                                    canvas.drawRect(box, paint)
                                 }
                             }
-
-                            val config =
-                                AppDialogConfig(getContext(), R.layout.barcode_result_dialog)
-                            config.setContent(buffer)
-                                .setHideCancel(true)
-                                .setOnClickConfirm {
-                                    AppDialog.INSTANCE.dismissDialog()
-                                }
-                            val imageView = config.getView<ImageView>(R.id.ivDialogContent)
-                            imageView.setImageBitmap(bitmap)
-                            AppDialog.INSTANCE.showDialog(config)
-                        } else {
-                            LogUtils.d("result is null")
-                            Toast.makeText(getContext(), "result is null", Toast.LENGTH_SHORT)
-                                .show()
                         }
+                        val config = AppDialogConfig(getContext(), R.layout.barcode_result_dialog)
+                        config.setContent(buffer)
+                            .setHideCancel(true)
+                            .setOnClickConfirm {
+                                AppDialog.INSTANCE.dismissDialog()
+                            }
+                        val imageView = config.getView<ImageView>(R.id.ivDialogContent)
+                        imageView.setImageBitmap(bitmap)
+                        AppDialog.INSTANCE.showDialog(config)
+                    } else {
+                        // 没有结果
+                        LogUtils.d("result is empty")
+                        showToast("result is empty")
                     }
-
-                    override fun onFailure(e: Exception?) {
-                        LogUtils.d("onFailure")
-                        Toast.makeText(getContext(), "onFailure", Toast.LENGTH_SHORT).show()
-                    }
-
-                    // 如果指定具体的识别条码类型，速度会更快
-                }, if (isQrCode) Barcode.FORMAT_QR_CODE else Barcode.FORMAT_ALL_FORMATS)
+                }.addOnFailureListener(this) { e ->
+                    // 失败
+                    LogUtils.w(e)
+                }
             } catch (e: Exception) {
-                e.printStackTrace()
-                Toast.makeText(getContext(), e.message, Toast.LENGTH_SHORT).show()
+                LogUtils.w(e)
             }
+        }
+    }
 
+    /**
+     * 根据Uri获取对应的图片
+     */
+    private fun Uri.getBitmap(): Bitmap {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            val source = ImageDecoder.createSource(contentResolver, this)
+            ImageDecoder.decodeBitmap(source)
+        } else {
+            MediaStore.Images.Media.getBitmap(contentResolver, this)
         }
     }
 
@@ -174,17 +190,18 @@ class MainActivity : AppCompatActivity() {
      * 选择照片 - 条形码/二维码 图片识别
      */
     private fun startPickPhoto() {
-        val pickIntent = Intent(
-            Intent.ACTION_PICK,
-            MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        )
+        val pickIntent = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
         pickIntent.setDataAndType(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, "image/*")
         startActivityForResult(pickIntent, REQUEST_CODE_PHOTO)
     }
 
     fun onClick(v: View) {
         when (v.id) {
-            R.id.btnQRCodeScanning -> startActivityForResult(Intent(this, QRCodeScanningActivity::class.java), REQUEST_CODE_SCAN_CODE)
+            R.id.btnQRCodeScanning -> startActivityForResult(
+                Intent(this, QRCodeScanningActivity::class.java),
+                REQUEST_CODE_SCAN_CODE
+            )
+
             R.id.btnMultipleQRCodeScanning -> startActivity(MultipleQRCodeScanningActivity::class.java)
             R.id.btnBarcodeScanning -> startActivity(BarcodeScanningActivity::class.java)
             R.id.btnQRCodeRecognitionFromImage -> pickPhotoClicked(true)