+++ app/src/main/java/com/king/mlkit/vision/app/pose/AccuratePoseDetectionActivity.kt
@@ -17,7 +17,8 @@ class AccuratePoseDetectionActivity : AccuratePoseCameraScanActivity() {
 
     override fun onScanResultCallback(result: AnalyzeResult<Pose>) {
         if (result.result.allPoseLandmarks.isNullOrEmpty()) {
-            // 过滤掉空数据
+            // 过滤掉空数据，继续分析
+            cameraScan.setAnalyzeImage(true)
             return
         }
         cameraScan.setAnalyzeImage(false)