+++ mlkit-common/src/main/java/com/king/mlkit/vision/common/analyze/CommonAnalyzer.java
@@ -17,20 +17,20 @@ package com.king.mlkit.vision.common.analyze;
 
 import android.graphics.ImageFormat;
 
+import androidx.annotation.NonNull;
+import androidx.camera.core.ImageProxy;
+
 import com.google.android.gms.tasks.Task;
 import com.google.mlkit.vision.common.InputImage;
 import com.king.camera.scan.AnalyzeResult;
 import com.king.camera.scan.FrameMetadata;
 import com.king.camera.scan.analyze.Analyzer;
+import com.king.camera.scan.util.ImageUtils;
 
-import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
-
-import androidx.annotation.NonNull;
-import androidx.camera.core.ImageProxy;
 
 /**
  * 通用分析器：将相机预览帧数据分析的通用业务进行统一处理，从而简化各子类的实现；（适用于MLKit各个字库）
@@ -55,7 +55,7 @@ public abstract class CommonAnalyzer<T> implements Analyzer<T> {
     protected abstract Task<T> detectInImage(InputImage inputImage);
 
     @Override
-    public void analyze(@NonNull ImageProxy imageProxy, @NonNull OnAnalyzeListener<AnalyzeResult<T>> listener) {
+    public void analyze(@NonNull ImageProxy imageProxy, @NonNull OnAnalyzeListener<T> listener) {
         if (!joinQueue.get()) {
             int imageSize = imageProxy.getWidth() * imageProxy.getHeight();
             byte[] bytes = new byte[imageSize + 2 * (imageSize / 4)];
@@ -67,7 +67,7 @@ public abstract class CommonAnalyzer<T> implements Analyzer<T> {
         }
         final byte[] nv21Data = queue.poll();
         try {
-            yuv_420_888toNv21(imageProxy, nv21Data);
+            ImageUtils.yuv_420_888toNv21(imageProxy, nv21Data);
             InputImage inputImage = InputImage.fromByteArray(
                     nv21Data,
                     imageProxy.getWidth(),
@@ -77,7 +77,7 @@ public abstract class CommonAnalyzer<T> implements Analyzer<T> {
             );
             // 检测分析
             detectInImage(inputImage).addOnSuccessListener(result -> {
-                if (result == null || (result instanceof Collection && ((Collection<?>) result).isEmpty())) {
+                if (isNullOrEmpty(result)) {
                     queue.add(nv21Data);
                     listener.onFailure(null);
                 } else {
@@ -99,56 +99,18 @@ public abstract class CommonAnalyzer<T> implements Analyzer<T> {
     }
 
     /**
-     * YUV420_888转NV21
+     * 是否为空
      *
-     * @param image
-     * @param nv21
+     * @param obj
+     * @return
      */
-    private void yuv_420_888toNv21(@NonNull ImageProxy image, byte[] nv21) {
-        ImageProxy.PlaneProxy yPlane = image.getPlanes()[0];
-        ImageProxy.PlaneProxy uPlane = image.getPlanes()[1];
-        ImageProxy.PlaneProxy vPlane = image.getPlanes()[2];
-
-        ByteBuffer yBuffer = yPlane.getBuffer();
-        ByteBuffer uBuffer = uPlane.getBuffer();
-        ByteBuffer vBuffer = vPlane.getBuffer();
-        yBuffer.rewind();
-        uBuffer.rewind();
-        vBuffer.rewind();
-
-        int ySize = yBuffer.remaining();
-
-        int position = 0;
-
-        // Add the full y buffer to the array. If rowStride > 1, some padding may be skipped.
-        for (int row = 0; row < image.getHeight(); row++) {
-            yBuffer.get(nv21, position, image.getWidth());
-            position += image.getWidth();
-            yBuffer.position(Math.min(ySize, yBuffer.position() - image.getWidth() + yPlane.getRowStride()));
+    private boolean isNullOrEmpty(Object obj) {
+        if (obj == null) {
+            return true;
         }
-
-        int chromaHeight = image.getHeight() / 2;
-        int chromaWidth = image.getWidth() / 2;
-        int vRowStride = vPlane.getRowStride();
-        int uRowStride = uPlane.getRowStride();
-        int vPixelStride = vPlane.getPixelStride();
-        int uPixelStride = uPlane.getPixelStride();
-
-        // Interleave the u and v frames, filling up the rest of the buffer. Use two line buffers to
-        // perform faster bulk gets from the byte buffers.
-        byte[] vLineBuffer = new byte[vRowStride];
-        byte[] uLineBuffer = new byte[uRowStride];
-        for (int row = 0; row < chromaHeight; row++) {
-            vBuffer.get(vLineBuffer, 0, Math.min(vRowStride, vBuffer.remaining()));
-            uBuffer.get(uLineBuffer, 0, Math.min(uRowStride, uBuffer.remaining()));
-            int vLineBufferPosition = 0;
-            int uLineBufferPosition = 0;
-            for (int col = 0; col < chromaWidth; col++) {
-                nv21[position++] = vLineBuffer[vLineBufferPosition];
-                nv21[position++] = uLineBuffer[uLineBufferPosition];
-                vLineBufferPosition += vPixelStride;
-                uLineBufferPosition += uPixelStride;
-            }
+        if (obj instanceof Collection) {
+            return ((Collection<?>) obj).isEmpty();
         }
+        return false;
     }
 }